aspe: An R Package to Analyse and Visualise River Fish Data in France

The aspe package provides a suite of tools for most of the common processing of the ASPE database doi:10.1051/kmae/2022021 , including importation of the tables from a dump of the PostGreSQL database, calculation of abundances, densities, size distributions, along with graphical output and spatial processing.

See also

Author

Maintainer: Pascal Irz pascal.irz@ofb.gouv.fr (ORCID)

Other contributors:

Cédric Mondy cedric.mondy@ofb.gouv.fr (ORCID) [contributor]

Benoît Richard benoit.richard@ofb.gouv.fr (ORCID) [contributor]

Lilian Bonnafoux l.bonnafoux@peche63.com [contributor]